Move buildSchedule function up to OrganizationEditPage.
